Move TriggersPreviewButton to a new component
